requirements.txt
torch
torchvision
opencv-python
matplotlib
trimesh
rembg
numpy
Pillow
